back-end/requirements.txt
flask
flask-cors
#YOLO packages
keras
numpy
Pillow
scipy
h5py
matplotlib
opencv-python
keras-resnet
tensorflow-macos
tensorflow-metal
protobuf==3.19


mysql-connector-python
python-dotenv
